==> notebooks com metodologia/1_preprocessamento_books.ipynb <==
import kagglehub
import os
import pandas as pd
import re
import nltk

# Download dataset latest version
path = kagglehub.dataset_download("mohamedbakhet/amazon-books-reviews")

print("Path to dataset files:", path)

# Listar os ficheiros no diretório descarregado
files = os.listdir(path)
print("Ficheiros disponíveis:", files)


# Caminhos completos para os ficheiros CSV
file1 = os.path.join(path, "books_data.csv") 
file2 = os.path.join(path, "Books_rating.csv")

# Carregar os CSVs em DataFrames
books_data = pd.read_csv(file1)
books_rating = pd.read_csv(file2)

# Verificar os primeiros registos
print(books_data.head())
print(books_rating.head())

# Explorar o dataset
books_rating.info()

# Explorar o dataset
books_data.info()

# Verificar missing values
NaN_books_data = books_data.isnull().sum()
print(NaN_books_data)

# Limpar dataset, excluindo missing values nas colunas relevantes
books_data_semNaN = books_data.dropna(subset=['Title', 'authors', 'categories', 'description'])

# Verificar novo dataset
books_data_semNaN.info()

# Excluir colunas irrelevantes do DF
books_data_semNaN.drop(columns=['previewLink', 'publisher', 'publishedDate', 'infoLink'], inplace=True)

# Verificar conteúdo do dataset limpo
books_data_semNaN.info()

# redução da amostra para 50% do dataset original, porque não estavamos a conseguir executar o código

# Realizar a amostragem estratificada manualmente
sample_booksdata = books_data_semNaN.groupby('categories', group_keys=False).apply(lambda x: x.sample(frac=0.5, random_state=42))

print(sample_booksdata)

# redução da amostra das reviews, com base na sample dos livros
sample_booksrating = books_rating[books_rating['Title'].isin(sample_booksdata['Title'])]

sample_booksrating.info()

# pré-processamento da sample de books data

sample_booksdata.info()

#lowercasing
sample_booksdata = sample_booksdata.map(lambda s: s.lower() if type(s) == str else s)

#punctuation removal
sample_booksdata = sample_booksdata.map(lambda x: re.sub(r'[^\w\s]', '', x) if isinstance(x, str) else x)  # Remove special characters
sample_booksdata = sample_booksdata.map(lambda x: re.sub(r'\d+', '', x) if isinstance(x, str) else x)  # Remove numbers
sample_booksdata = sample_booksdata.map(lambda x: re.sub(r'º', '', x) if isinstance(x, str) else x)  # Remove the "º" character
sample_booksdata = sample_booksdata.map(lambda x: ' '.join(x.split()) if isinstance(x, str) else x)  # Remove extra spaces

# remove stopwords
from nltk.corpus import stopwords

nltk.download('stopwords')

# Definir a lista de stopwords
stop_words = set(stopwords.words('english'))

# Função para remover stopwords
def remove_stopwords(text):
    words = text.split()
    cleaned_text = ' '.join([word for word in words if word.lower() not in stop_words])
    return cleaned_text

# Converter os valores das colunas para string antes de aplicar a função
sample_booksdata['description'] = sample_booksdata['description'].astype(str).apply(remove_stopwords)

print(sample_booksdata)

# Remove words that are shorter than or equal to 2 characters
# Função para remover palavras curtas
def remove_short_words(text):
    return ' '.join([word for word in text.split() if len(word) > 2])

# Aplicar a função de remoção de palavras curtas à coluna
sample_booksdata['description'] = sample_booksdata['description'].apply(remove_short_words)

print(sample_booksdata)

#Devolver a coluna original de authors à sample_booksdata
sample_booksdata['authors'] = books_data['authors']

print(sample_booksdata['authors'])

sample_booksdata.drop(columns=['ratingsCount'], inplace=True)
sample_booksdata.head()

#Devolver a coluna original de titulo à sample_booksdata
sample_booksdata['title_raw'] = books_data['Title']

print(sample_booksdata['title_raw'])

# pré-processamento da sample de book's ratings

sample_booksrating.info()

NaN_books_rating = sample_booksrating.isnull().sum()
print(NaN_books_rating)

# dropar as linhas com NaN
sample_booksrating = sample_booksrating.dropna(subset=['review/summary', 'review/text'])

#dropar colunas desnecessárias
sample_booksrating.drop(columns=['Price', 'review/helpfulness'], inplace=True)

sample_booksrating.info()

#lowercasing
sample_booksrating = sample_booksrating.map(lambda s: s.lower() if type(s) == str else s)

#ponctuation removal
sample_booksrating = sample_booksrating.map(lambda x: re.sub(r'[^\w\s]', '', x) if isinstance(x, str) else x)  # Remove special characters
sample_booksrating = sample_booksrating.map(lambda x: re.sub(r'\d+', '', x) if isinstance(x, str) else x)  # Remove numbers
sample_booksrating = sample_booksrating.map(lambda x: re.sub(r'º', '', x) if isinstance(x, str) else x)  # Remove the "º" character
sample_booksrating = sample_booksrating.map(lambda x: ' '.join(x.split()) if isinstance(x, str) else x)  # Remove extra spaces

# remove stopwords
from nltk.corpus import stopwords

nltk.download('stopwords')

# Definir a lista de stopwords
stop_words = set(stopwords.words('english'))

# Função para remover stopwords
def remove_stopwords(text):
    words = text.split()
    cleaned_text = ' '.join([word for word in words if word.lower() not in stop_words])
    return cleaned_text

# Converter os valores das colunas para string antes de aplicar a função
sample_booksrating['review/summary'] = sample_booksrating['review/summary'].astype(str).apply(remove_stopwords)
sample_booksrating['review/text'] = sample_booksrating['review/text'].astype(str).apply(remove_stopwords)

print(sample_booksrating)

# Remove words that are shorter than or equal to 2 characters
# Função para remover palavras curtas
def remove_short_words(text):
	return ' '.join([word for word in text.split() if len(word) > 2])

# Aplicar a função de remoção de palavras curtas às colunas
sample_booksrating['review/summary'] = sample_booksrating['review/summary'].apply(remove_short_words)
sample_booksrating['review/text'] = sample_booksrating['review/text'].apply(remove_short_words)

print(sample_booksrating.head())

#devolver a coluna original de titulo à sample_booksrating
sample_booksrating['title_raw'] = books_rating['Title']

sample_booksrating.head()

# salvar amostras num csv

#verificar diretório de trabalho
print(os.getcwd())

sample_booksdata.to_csv('sample_booksdata.csv', index=False)
sample_booksrating.to_csv('sample_booksrating.csv', index=False)
