==> notebooks com metodologia/2_filtragem_genres.ipynb <==
import pandas as pd
import os

#definir diretório
os.chdir(f'C:\PAC')

#carregar dataset
books_data = pd.read_csv('sample_booksdata.csv')

books_data.info()

category_counts = books_data['categories'].value_counts()
print(category_counts)

top_100_categories = category_counts.head(100)
print(top_100_categories)

top_100_categories.to_csv('top_100_categories.csv')

top_100_categories_df = top_100_categories.reset_index()
filtered_books_data = books_data[books_data['categories'].isin(top_100_categories_df['categories'])]

filtered_books_data.info()

# Importar ficheiro com Thesaurus relacionando categorias e gêneros
thesaurus_genres = pd.read_csv('thesaurus_genres.csv', sep=';')

thesaurus_genres.info()

# Atribuir genre a cada um dos livros, com base nas categorias
books_with_genres = filtered_books_data.merge(thesaurus_genres, on='categories', how='left')
books_with_genres['genres'] = books_with_genres['count']


print(books_with_genres.info())

books_with_genres['genres'].unique()

# Verificar ficheiro resultante
books_with_genres.head(5)

# Dropar colunas irrelevantes
books_with_genres.drop(columns=['count'], inplace=True)

# Salvar DF para um CSV
books_with_genres.to_csv('books_with_genres.csv', index=False)
